+++ coda/linalg/Vector.h
@@ -27,47 +27,47 @@ public:
     uint size() const;
 
     const Vector& operator=(const Vector& v);
-    const Vector& operator=(float a);
+    const Vector& operator=(double a);
 
     void zeros();
 
     void abs();
 
-    void add(float a);
-    void add(float a, Vector& v);
-    void add(float a, Vector&v, float b, Vector& w);
+    void add(double a);
+    void add(double a, Vector& v);
+    void add(double a, Vector&v, double b, Vector& w);
 
-    void eq(float a);
-    void eq(float a, Vector& v);
-    void eq(float a, Vector&v, float b, Vector& w);
+    void eq(double a);
+    void eq(double a, Vector& v);
+    void eq(double a, Vector&v, double b, Vector& w);
 
-    void axpy(float a, const Vector& v);
+    void axpy(double a, const Vector& v);
 
-    float dot(const Vector& x) const;
-    float norm(std::string norm_type) const;
+    double dot(const Vector& x) const;
+    double norm(std::string norm_type) const;
 
-    float min() const;
-    float max() const;
-    float sum() const;
+    double min() const;
+    double max() const;
+    double sum() const;
 
     std::string str() const;
 		void print() const;
 
-    const Vector& operator*=(float a);
-    const Vector& operator/=(float a);
-    const Vector& operator+=(float a);
-    const Vector& operator-=(float a);
+    const Vector& operator*=(double a);
+    const Vector& operator/=(double a);
+    const Vector& operator+=(double a);
+    const Vector& operator-=(double a);
     const Vector& operator+=(const Vector& v);
     const Vector& operator-=(const Vector& v);
 
 		void matvec(Matrix& A, Vector& y);
 
-    inline float operator[] (const uint index) const
+    inline double operator[] (const uint index) const
     {
         return _values[index];
     };
 
-    inline float& operator[] (const uint index)
+    inline double& operator[] (const uint index)
     {
         return _values[index];
     };
@@ -94,10 +94,10 @@ public:
 
 private:
     uint _size;
-    float* _values;
+    double* _values;
 
 private:
-    float* memptr();
+    double* memptr();
 };
